# www/api/openapi.yaml
openapi: 3.0.3
info:
  title: RootWarden API
  description: |
    API REST du backend Flask de RootWarden — plateforme de gestion centralisee
    de serveurs Linux (cles SSH, mises a jour, iptables, scan CVE).

    **Authentification** : header `X-API-KEY` obligatoire sur toutes les routes (sauf /test).
    Le frontend PHP utilise `/api_proxy.php` qui injecte la cle automatiquement.
  version: 1.11.0
  contact:
    name: RootWarden
    url: https://github.com/Timikana/rootwarden
servers:
  - url: https://localhost:5000
    description: Backend direct (dev)
  - url: /api_proxy.php
    description: Via proxy PHP (production)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    TempPermission:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        user_name: { type: string }
        permission: { type: string }
        machine_id: { type: integer, nullable: true }
        granted_by_name: { type: string }
        reason: { type: string }
        expires_at: { type: string, format: date-time }
security:
  - ApiKeyAuth: []
paths:
...
  /admin/temp_permissions:
    get:
      tags: [Admin]
      summary: Liste des permissions temporaires actives
      responses:
        '200':
          description: Permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  permissions:
                    type: array
                    items: { $ref: '#/components/schemas/TempPermission' }
    post:
      tags: [Admin]
      summary: Accorder une permission temporaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, permission]
              properties:
                user_id: { type: integer }
                permission: { type: string, example: "can_deploy_keys" }
                hours: { type: integer, minimum: 1, maximum: 720, default: 24 }
                machine_id: { type: integer, nullable: true }
                reason: { type: string }
      responses:
        '200': { description: Permission accordee }
        '400': { description: Champs manquants ou duree invalide }

  # ── Fail2ban ──────────────────────────────────────────────────────────────
...
  /exclude_user:
    post:
      tags: [Admin]
      summary: Exclure un utilisateur d'un serveur
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [user_id, machine_id]
              properties:
                user_id: { type: integer }
                machine_id: { type: integer }
      responses:
        '200': { description: Exclusion appliquee }

  # ── Admin ────────────────────────────────────────────────────────────────
  /server_lifecycle:
    post:
      tags: [Admin]
      summary: Change le statut lifecycle d'un serveur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, lifecycle_status]
              properties:
                machine_id: { type: integer }
                lifecycle_status: { type: string, enum: [active, retiring, archived] }
                retire_date: { type: string, format: date }
      responses:
        '200': { description: Mis a jour }
        '400': { description: Parametres invalides }